fold desktop/mobile separation back together
